used proper architecture methodology
