--- src/routes/index.routes.js
@@ -1,8 +1,0 @@
-const router = require('express').Router();
-const userRouter = require("./users.routes.js");
-
-router.use("/users", userRouter);
-
-
-
-module.exports = router;
